ai-backend/requirements.txt
flask
tensorflow
pillow
flask-cors
scipy